drm/nouveau/core: import ioctl/event interfaces
This forms the basis for the new APIs that will be exposed to userspace,
giving it access to:

- Object method calls, the immediately useful of which is performance
  counters and the abiity to manipulate the ZBC tables.
- Information on the child classes an object supports, in order to avoid
  having to try all supported classes until successful.
- Notifications, which will be used in the future to inform the client
  if its channel was killed due to a lockup, etc.

This commit imports the interfaces, but are not currently used.  The DRM
portion of the driver will be ported to speak to the core using these
interfaces as much as possible.

+++ drivers/gpu/drm/nouveau/core/include/core/notify.h
@@ -15,7 +15,7 @@ struct nvkm_notify {
 	/* set by nvkm_event ctor */
 	u32 types;
 	int index;
-	u8  size;
+	u32 size;
 
 	struct work_struct work;
 	/* this is const for a *very* good reason - the data might be on the
